Plugin headers: require specific PHP and WP versions
WordPress Core includes a mechanism to block plugin activations when your site's version of PHP or WP does match the requirements of the plugin.
This was implemented as a result of this trac ticket:
https://core.trac.wordpress.org/ticket/43992

This ensures that site owners do not find themselves with Fatal errors on their site after activating a plugin.